'商品管理’

+++ src/views/home/Home.vue
@@ -2,30 +2,32 @@
   <el-container class="container">
     <el-header class="header">
       <el-row>
-        <el-col :span="22">
+        <el-col :span="23">
           <div class="grid-content bg-purple">
             <img src="~assets/img/logo.png" alt="">
             <span class="text">后台管理系统</span>
           </div>
         </el-col>
-        <el-col :span="2">
+        <el-col :span="1">
           <el-row>
-            <el-button type="danger"  @click.prevent="logout">sign up</el-button>
+<!--            <el-button type="danger"  @click.prevent="logout">sign up</el-button>-->
+            <span class="out"><i class="el-icon-switch-button" @click.prevent="logout"></i></span>
           </el-row>
         </el-col>
       </el-row>
     </el-header>
-    <el-container>
-      <el-aside width="200px" class="aside">
-      <!--侧边导航-->
+    
+    <el-container class="aside-menu">
+      <el-aside :width="isCollapse ? '64px' : '180px'" class="aside">
+        <div class="toggle-button" @click="toggleMenu"><i class="el-icon-d-arrow-right"></i></div>
         <el-menu
           default-active="2"
-          class="el-menu-vertical-demo"
-          background-color="#333744"
           text-color="#ffffff"
+          active-text-color="gold"
           :unique-opened="true"
-          :router="true">
-          
+          :router="true"
+          :collapse="isCollapse"
+          >
           <!--左侧菜单-->
           <el-submenu  v-for="(item,index) in menuList" :index="item.id+''" :key="item.id">
             <template slot="title">
@@ -40,6 +42,7 @@
             </el-menu-item>
           </el-submenu>
         </el-menu>
+      <!--侧边导航-->
       </el-aside>
       <el-main class="main">
         <router-view/>
@@ -55,23 +58,25 @@
 			return {
 		    //左侧菜单图标
 				iconObj: {
-					'125': 'el-icon-user',
-					'103': 'el-icon-folder',
-					'101': 'el-icon-shopping-bag-2',
-					'102': 'el-icon-shopping-cart-2',
-					'145': 'el-icon-s-data'
+					'125': 'el-icon-user-solid',
+					'103': 'el-icon-s-tools',
+					'101': 'el-icon-s-goods',
+					'102': 'el-icon-s-shop',
+					'145': 'el-icon-s-marketing'
 				},
         //默认路径
         activePath:'',
-        menuList:[]
+        menuList:[],
+				isCollapse:true,
+        width:''
       }
     },
 		created() {
-			//验证token,不存在则跳转到登录页面
-			this.activePath = window.sessionStorage.getItem('token')
-			if(!this.activePath){
-				this.$router.push('/')
-			}
+			// //验证token,不存在则跳转到登录页面
+			// this.activePath = window.sessionStorage.getItem('token')
+			// if(!this.activePath){
+			// 	this.$router.push('/')
+			// }
 			this.getMenuList()
 		},
     methods:{
@@ -83,10 +88,19 @@
 			},
       //登出
       logout(){
-				this.$message.success('登出成功')
-				//清除token
-				window.sessionStorage.clear()
-        this.$router.push('/')
+				this.$confirm('是否退出登录', '提示', {
+					confirmButtonText: '确定',
+					cancelButtonText: '取消',
+					type: 'warning'
+				}).then(() => {
+					this.$message.success('退出成功')
+					//清除token
+					window.sessionStorage.clear()
+					this.$router.push('/')
+				})
+      },
+			toggleMenu(){
+				this.isCollapse = !this.isCollapse
       }
     }
 	}
@@ -97,22 +111,22 @@
     height: 100%;
   }
   .header {
-    background-color: #373f41;
+    background-color: #2d3741;
     color: #333;
     line-height: 60px;
     justify-content:space-between;
     align-items: center;
   }
   .aside {
-    background-color: #333744;
-    color: #333;
-    line-height: 200px;
+    color: #eeeeee;
+    background: #333744;
   }
-  
+  .el-menu{
+    background: #333744;
+  }
   .main {
-    background-color:  #eaedf1;
-    color: #333;
-    text-align: center;
+    background-color:#eaedf1;
+    color: #353f36;
   }
   .header img{
     height: 50px;
@@ -126,11 +140,19 @@
     font-size: 20px;
     color: #ddfffc;
   }
-  el-button{
-    /*鼠标变小手*/
+  .toggle-button {
+    background-color: #4A5064;
+    font-size: 16px;
+    line-height:24px;
+    color: darkseagreen;
+    text-align: center;
+    letter-spacing: 0.2em;
     cursor: pointer;
   }
-  .el-menu-vertical-demo{
-   width: 100%;
+  .out{
+    /*鼠标变小手*/
+    cursor: pointer;
+    font-size: 24px;
+    color: red;
   }
 </style>
